style: migrate to real use of Optional

+++ src/main/java/org/esupportail/publisher/web/rest/ItemResource.java
@@ -132,7 +132,7 @@ public class ItemResource {
     public ResponseEntity<?> update(@RequestBody ActionDTO action) throws URISyntaxException {
         log.debug("REST request to update Item with action : {}", action);
         Optional<AbstractItem> optionalAbstractItem =  itemRepository.findById(action.getObjectId());
-        AbstractItem item = optionalAbstractItem == null || !optionalAbstractItem.isPresent()? null : optionalAbstractItem.get();
+        AbstractItem item = optionalAbstractItem.orElse(null);
         switch(action.getAttribute()) {
             case "enclosure" :
                 return contentService.setEnclosureItem(action.getValue(), item);
@@ -183,7 +183,7 @@ public class ItemResource {
     public ResponseEntity<AbstractItem> get(@PathVariable Long id, HttpServletResponse response) {
         log.debug("REST request to get Item : {}", id);
         Optional<AbstractItem> optionalAbstractItem =  itemRepository.findById(id);
-        AbstractItem item = optionalAbstractItem == null || !optionalAbstractItem.isPresent()? null : optionalAbstractItem.get();
+        AbstractItem item = optionalAbstractItem.orElse(null);
         if (item == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -203,7 +203,11 @@ public class ItemResource {
     public void delete(@PathVariable Long id) {
         log.debug("REST request to delete Item : {}", id);
         Optional<AbstractItem> optionalAbstractItem =  itemRepository.findById(id);
-        AbstractItem item = optionalAbstractItem == null || !optionalAbstractItem.isPresent()? null : optionalAbstractItem.get();
+        AbstractItem item = optionalAbstractItem.orElse(null);
+        if (item == null) {
+            log.warn("Try to delete an item not existing !");
+            return;
+        }
         fileService.deleteInternalResource(item.getEnclosure());
         itemRepository.deleteById(id);
     }
