+++ src/main/java/org/esupportail/publisher/web/rest/NewsResource.java
@@ -15,12 +15,21 @@
  */
 package org.esupportail.publisher.web.rest;
 
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Optional;
+
+import javax.inject.Inject;
+import javax.servlet.http.HttpServletResponse;
+
 import org.esupportail.publisher.domain.AbstractItem;
 import org.esupportail.publisher.domain.News;
 import org.esupportail.publisher.repository.NewsRepository;
 import org.esupportail.publisher.security.SecurityConstants;
 import org.esupportail.publisher.service.FileService;
 import org.esupportail.publisher.web.rest.util.PaginationUtil;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -29,14 +38,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
-
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletResponse;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 /**
  * REST controller for managing News.
@@ -110,7 +117,7 @@ public class NewsResource {
     public ResponseEntity<News> get(@PathVariable Long id, HttpServletResponse response) {
         log.debug("REST request to get News : {}", id);
         Optional<News> optionalNews =  newsRepository.findById(id);
-        News news = optionalNews == null || !optionalNews.isPresent()? null : optionalNews.get();
+        News news = optionalNews.orElse(null);
         if (news == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -127,7 +134,11 @@ public class NewsResource {
     public void delete(@PathVariable Long id) {
         log.debug("REST request to delete News : {}", id);
         Optional<News> optionalNews =  newsRepository.findById(id);
-        AbstractItem item = optionalNews == null || !optionalNews.isPresent()? null : optionalNews.get();
+        AbstractItem item = optionalNews.orElse(null);
+        if (item == null) {
+            log.warn("Try to delete an item not existing !");
+            return;
+        }
         fileService.deleteInternalResource(item.getEnclosure());
         newsRepository.deleteById(id);
     }
